requirements.txt
# Core Flask Application
Flask>=3.0.0
Flask-RESTful>=0.3.10
Flask-CORS>=4.0.0
python-dotenv>=1.0.0

# PDF Processing
pymupdf<1.25.3

# AI/ML - Google Gemini
google-generativeai>=0.3.0

# Mathematical Solving
sympy>=1.13.0

# Document Generation
python-docx>=1.1.0

# Database (Optional - for metadata storage)
pymongo>=4.0.0

# pdf2zh_next Essential Dependencies
# Note: pdf2zh_next is a local package (install with: pip install -e .)
# These are the core dependencies required for PDF translation:
babeldoc>=0.5.7,<0.6.0
pydantic>=2.10.6
pydantic-settings>=2.8.1
httpx>=0.28.1
requests>=2.31.0
tqdm>=4.66.0
tenacity>=8.2.0
numpy>=1.24.0
fontTools>=4.0.0
rich>=13.0.0
chardet>=5.2.0
pyyaml>=6.0.2
